refactor: ManualHandlerAdapter 모듈 위치 변경
