+++ mvc/src/main/java/nextstep/mvc/controller/asis/ManualHandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package nextstep.mvc.controller.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
